requirements.txt
fastapi[all]
pydantic
PyYAML
requests
starlette
uvicorn
uvloop
aiohttp
async-timeout
starlette_exporter
paramiko
sentry-sdk
click
